+++ application/languages/fr/js.php
@@ -31,5 +31,6 @@ return [
     'guest_editor' => 'Rédacteur invité',
     'epiadmin' => 'Root',
     'other' => 'autre',
-    'not-assigned' => 'Non assigné'
+    'not-assigned' => 'Non assigné',
+    'orcid-duplicate' => 'Chaque ORCID doit être unique.'
 ];